feedback_analyzer/requirements.txt
streamlit
torch
transformers
openai
python-dotenv
pandas
nltk
scikit-learn
accelerate
janome
pyngrok